feat(translations): KOR-60: Moved scrapeTranslations.ts script outside of src
